# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApnaFarm API",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/rewards": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get the rewards for the authenticated user",
                "operationId": "0e90644a33fd2077de567b3d46696d5a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "total_rewards": {
                                                    "type": "string"
                                                },
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "amount": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "total_rewards": {
                                                    "type": "string",
                                                    "example": "0.00"
                                                },
                                                "amount": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "amount": {
                                                                "type": "string",
                                                                "example": "0.00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No rewards found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Add or update the rewards for the authenticated user",
                "operationId": "bd95a311a20ad705f262a4670b2793a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "total_rewards": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cashback": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "referrals": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "promotions": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rewards updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "reward": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "total_rewards": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "cashback": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "referrals": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "promotions": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "wallet": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "cash_balance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "rewards_balance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/rewards/history": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get the rewards history for the authenticated user",
                "operationId": "f9d149fc50aadc3d0a08d2ec7600516e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "total_rewards": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "cashback": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "referrals": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "promotions": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No rewards history found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter your Bearer token in the format **Bearer &lt;token&gt;**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "Bearer": []
        }
    ]
}